backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
PyJWT==2.8.0

# HTTP and API
httpx==0.25.2
requests==2.31.0

# Data validation and serialization
pydantic==2.9.0
pydantic[email]==2.9.0
pydantic-settings==2.1.0
email-validator==2.2.0

# Environment and configuration
python-dotenv==1.0.0

# CORS middleware
fastapi-cors==0.0.6

# Code execution and AI services
google-generativeai==0.3.2
#judge0-api==1.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Monitoring and logging
structlog==23.2.0

# File handling
python-magic==0.4.27

# Date and time utilities
python-dateutil==2.8.2

# JSON handling
orjson==3.9.10

# Background tasks
celery==5.3.4
redis==5.0.1

# Email services
fastapi-mail==1.4.1

# File upload and processing
aiofiles==23.2.1
pandas==2.2.2
openpyxl==3.1.2

# Rate limiting
slowapi==0.1.9

# Health checks
psutil==5.9.6